Jira ticket: https://jira.huit.harvard.edu/browse/TLT-1920
updated bounce-back message to inform users that their message might have been undelivered if they are sending mail from an email account that is not their default email address in Canvas:

changed this:

"You have insufficient permissions to post to this list. Please contact the teaching staff of this course if you think this was in error."

to this:

"Either you have insufficient permission to email this list, or you didn’t set your current email address as your default address in Canvas. Please contact the teaching staff of this course for assistance."