exit regression suite with error on test failure
this change also bumps the versions of selenium and selenium_common,
with corresponding fixes to settings and some selenium tests to attempt
to get the tests to pass.  at this point they're still failing, but due
to a nonsense-looking tool_id being used.  since this change was just
supposed to get failures to return an error code, and they're now doing
that, i'm calling it a day here.